timezone: Fix use-after-free when cancelling GeoClue lookup
When rebuilding the timezone page after the user has chosen a language,
any ongoing GeoClue construction operation is cancelled. However, the
callback for that might be called after the page has been disposed
(since the operation doesn’t hold a strong ref on the page). Avoid
writing to `priv->geoclue_simple` in that case, until after checking the
error code.

Signed-off-by: Philip Withnall <withnall@endlessm.com>