gis-driver: Don't defer window state changes to idle handlers
gnome-initial-setup tries to conserve space on small screens by
maximizing the window. When it decides to maximize the window,
it does this unnecessarly from an idle handler.

Unfortunately, a bug somewhere else in the stack gets exposed by
using this idle handler.  Sometimes, the maximization request
doesn't make it through.

This commit cleans up the unnecessary idle handler, and at the same
time side steps the other bug.

https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/issues/110