data: Leave to meson the duty of keeping the lists in sync
Don't leave to humans what machines can do, the gnome-initial-setup.session
list is filled with an hard-to-maintain single-line semicolon separated
list of components, that then are almost repeated in the gnome systemd
session.conf.
And these are meant to be kept in sync.

Meson can help us in this, so let's just keep a single list in meson and
use it to generate the various formats we need.