account: Always define all callbacks mentioned in .ui
Previously, webcam_icon_selected was only defined if gnome-initial-setup
was built with Cheese support. However, it is always mentioned in the
.ui file. This leads to the following warning at runtime:

    Could not find signal handler 'webcam_icon_selected'.  Did you compile with -rdynamic?

Define a dummy callback in the no-Cheese case. Warn if it is called.
(The button should always be hidden if Cheese is not available, so it
should never be called.)

Fixes #114.