data: Install .desktop file to normal directory
When run for an existing user, Initial Setup currently runs within that
user's normal session. This means that Shell shows an icon for the app
in the top bar, dash, and Alt-Tab overlay.

However, previously the .desktop file was only installed to the GDM
greeter's own desktop file directory, meaning that it was not found by
the normal Shell session, and a missing-icon icon would be shown
instead.

Previously, some distro packagers have worked around this by symlinking
the .desktop file to the normal location. But the greeter-specific
directory is added to the greeter's XDG_DATA_DIRS, rather than replacing
it; so we can simply install the desktop file to the normal location.
NoDisplay=true ensures it won't be shown to the user unless it is
launched manually or by some other means.

Fixes #52