Disable existing-user mode
Prior to GNOME 40, GNOME Tour was launched at the end of Initial Setup's
existing-user mode. In GNOME 40, Shell displays a dialog offering to run
the tour (for existing users which have not logged in, and users
upgrading to 40). If we also run Initial Setup for existing users who
have not previously logged in, they would then end up with Tour on top
of Initial Setup, which would not be a good look.

Ideally, we would like to always run Initial Setup in the "kiosk" mode
that is used when no users exist. (The lack of this means that, for
example, the language-selection page doesn't actually work until the
user restarts their session.) This would also avoid the collision with
Tour. Until we have this, disable existing-user mode.

One option would have been to remove the xdg-autostart and systemd units
for this mode. The downside of doing that would be that the
.config/gnome-initial-setup-done file would not be written, so if we
reinstate this mode in a future release, we wouldn't be able to
distinguish "user has never logged in" from "user has logged in, but
with a GNOME version that didn't do existing-user mode Initial Setup".

See: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/issues/120#note_1019004>
See: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/issues/12>
Part-of: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/113>