timezone: fix critical on startup when geoclue is disabled
gnome-initial-setup is crashing on startup with this error:

(gnome-initial-setup:85469): InitialSetup-CRITICAL **: 09:18:43.280: Failed to connect to GeoClue2 service: GDBus.Error:org.freedesktop.DBus.Error.AccessDenied: Geolocation disabled for UID 1000

g_critical is only to be used to indicate programmer error. In this
case, it's entirely expected that we might fail to connect to geoclue,
so even g_warning() would be too much. I've changed the error message to
use g_info() instead, which does not print by default, though we could
just as well get rid of it entirely.