software: replace use of obsolete settings with fedora-third-party
The original repo enablement code relies on two GNOME Software settings
that no longer exist: one setting to toggle whether properietary repos
are enabled, and another that lists which proprietary repos to enable.
gnome-initial-setup would then use PackageKit to enable said repos. This
was nice in that it worked using entirely upstream code, but since it no
longer exists that does us no good.

Instead we'll just have distro-specific implementation here, which is a
little non-ideal but not the end of the world. We'll just call a
Fedora-specific script depending on whether the setting is enabled or
disabled. Feel free to add logic for your distro.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/121>