Add polkit rule to allow executing fedora-third-party
This is based on an Endless commit by Philip Chimento.

Note that gnome-initial-setup will execute fedora-third-party wherever
it is found in PATH, since there is some debugging benefit to allowing
this, but it will require a password prompt unless it's the trusted
version of the script installed in bindir.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/121>