Standardize COPYING to just contain GPLv2
For reasons unknown to me, COPYING previously contained the full text of
LGPLv2 (not 2.1!) despite the entirety of this project being under GPL
v2 or later. It also contained a preamble that the license is >= 2, not
== 2.

Make COPYING a verbatim copy of GPL v2, and add the "or later" note to
README. This appears to be standard for similarly-licensed GNOME
projects I checked (including gnome-shell and mutter). This will also
cause GitLab to display the correct license, rather than "This project
is licensed under the Other." with a broken link.