goa: Check connectivity, not presence of a default route
g_network_monitor_get_network_available() just indicates that there is a
default route for at least one of IPv4 or IPv6, not that the wider
internet can be reached. If you are behind a captive portal, or are on a
network not connected to the wider internet, then you will not in fact
be able to log into your Google account.

Instead, monitor the connectivity property, which gives a more nuanced
view of the situation.