Ask systemd to be respawned on failure
Currently, if Initial Setup crashes, you are left staring at a blank
wallpaper, with no option but to shut down the computer.

If the session is managed by systemd, we can ask it to respawn
gnome-initial-setup.service (this application) on failure. The user will
be dumped back at the first page, which is not ideal, but is still an
improvement. In addition, certain settings (like locale, keyboard
layout, Wi-Fi network, and configured online accounts) are applied at
the point they are chosen, so will be preserved across a crash and
relaunch.

Fixes #26