Tell libhandy not to use Settings portal
By default, libhandy reads settings from the Settings portal, which causes
the portal to be started, which causes gnome-keyring to be started. This
interferes with our attempt to manually start gnome-keyring and set the
login keyring password to a well-known value, which we overwrite with
the user's password once they choose one.

To avoid this problem, set an environment variable to tell libhandy not
to use the portal.

Fixes: https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/issues/141