um-utils: Remove incorrect NULL check
g_utf8_next_char(p) returns a pointer to the next UTF-8 character in the
given string. This is always p + n for a small constant n.

Assuming p points to a well-formed, 0-byte-terminated UTF-8 string, p +
n is always non-NULL. And GCC warns about this:

    ../gnome-initial-setup/pages/account/um-utils.c: In function 'extract_initials_from_name':
    ../gnome-initial-setup/pages/account/um-utils.c:520:47: error: the comparison will always evaluate as 'true' for the pointer operand in 'p + (sizetype)*(g_utf8_skip + (sizetype)*(const guchar *)p)' must not be NULL [-Werror=address]
      520 |         if (p != NULL && g_utf8_next_char (p) != NULL) {

I think the intention here is to check if *(g_utf8_next_char(p)) is not
(char) 0. This is guaranteed by the use of g_strstrip() above: the
string does not end with whitespace. But it also wouldn't matter if it
were: g_string_append_unichar (initials, 0) would just append a
redundant NUL terminator.

Remove the unnecessary check.