account: Port to GTK4
Boy what a pain this one was.

A lot needed to be updated, but most of the complexity of this
page was in the GtkGrid updates, which requires switching the
UI files from <packing> to <layout>.

Use GtkEventControllerFocus to track focus out events.

Switch a password entries to GtkPasswordEntry, and port a ton of
GtkEntry calls to new GtkEditable API.