Smoke-test gnome-initial-setup --help
Initial Setup is currently crashing on startup in GNOME OS with:

    (process:2603): libsoup-ERROR **: 07:05:13.282: libsoup2 symbols
    detected. Using libsoup2 and libsoup3 in the same process is not
    supported.

This is because Initial Setup was built against soup3 and with parental
controls enabled; unfortunately, malcontent-ui links to libflatpak and
libostree which are currently built with soup2.

Similar scenarios may occur in the future. In lieu of an actual test
suite, test that 'gnome-initial-setup --help' succeeds.

https://gitlab.gnome.org/GNOME/gnome-build-meta/-/issues/513