accounts: Also check against NAME_REGEX if provided
In addition to checking if each character of the proposed username
conforms to UNIX concentions, if /etc/adduser.conf is present and
it provides NAME_REGEX, try and match the username against this
regex too.

This should fix cases where Initial Setup thinks the password is
okay, but accountsservice rejects the proposed username.

To make error reporting decent, check if the provided regex is the
lowercase-only regex, which is probably the most common one.