account: Fix buttons in enterprise login dialog
Since 13de947f4ff3444e7085007ff353e97bd74832ae ("account: Port to GTK4")
the label on the cancel button was lost, as was the use_underline property
for the continue button.

gtk-cancel is no more, but happily the translatable string "_Cancel" is
already present elsewhere in this project.

While we are here, mark the Continue button as the suggested action.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/159>