account: Set enterprise password errors as tooltip
In GNOME 42 and earlier, this error message was used as the tooltip for
an error icon in the entry. Since the GTK 4 port, the password widget is
now a GtkPasswordEntry, which does not support adding an inline icon in
this way.

e495fb58 ("account: Fix criticals when validating domain in enterprise
login dialog") stopped attempting to add an icon using the GtkEntry
functions which don't work on GtkPasswordEntry, making this page
work without warnings/criticals once again, but in the process the error
message would not be shown anywhere.

Set the error message (from realmd or the server, I'm not sure!) as the
tooltip for the entire entry, so that it is at least accessible. Clear
it when clearing the error state.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/161>