account: Install CSS in class_init(), not init()
GisAccountPageLocal may be instantiated multiple times, because all
pages are reconstructed when the language is changed. But the style
context is global (to a GdkDisplay). So previously, the same style
provider would be installed every time the language is changed.

Instead, do this in class_init().

Fixes 0416c80024e1c783df010b293173e919de7ceaa3 ("gis-account-page: Add a proper edit button to customize the user avatar").