Add helper function to add CSS from a resource
This project keeps a number of CSS files next to the widgets they apply
to. (You might expect from the current layout that they apply only to
the page at hand, but in fact they are global.)

Libadwaita can automatically load certain resources, but not when
arranged in this way. It may in future be desirable to centralize all
CSS rules in one file, and namespace them better.

https://gnome.pages.gitlab.gnome.org/libadwaita/doc/main/styles-and-appearance.html#custom-styles

In the meantime, add a helper function to load CSS from a resource and
install it for the default display.