driver: Set all categories when changing locale
Previously, Initial Setup would only change the LC_MESSAGES locale
category when the user selects a different locale on the language page.
However, this caused non-ASCII characters to be mangled after switching
locale. The following test program demonstrates the problem:

    #include "config.h"
    #include <locale.h>
    #include <glib.h>
    #include <glib/gi18n-lib.h>
    #include <libgweather/gweather.h>

    gint
    main (gint   argc,
          gchar *argv[])
    {
      /* gettext boilerplate */
      setlocale (LC_ALL, "");
      bindtextdomain (GETTEXT_PACKAGE, GNOMELOCALEDIR);
      bind_textdomain_codeset (GETTEXT_PACKAGE, "UTF-8");
      textdomain (GETTEXT_PACKAGE);

      /* Change locale to Czech */
      locale_t locale = newlocale (LC_MESSAGES_MASK, "cs_CZ.utf8", (locale_t) 0);
      g_assert (locale != (locale_t) 0);
      uselocale (locale);

      GWeatherLocation *world = gweather_location_get_world ();
      GWeatherLocation *praha = gweather_location_find_nearest_city (world, 50.08804, 14.42076);
      g_message ("%s", gweather_location_get_name (gweather_location_get_parent (praha)));

      return 0;
    }

The desired output is “Česká republika”, but this program instead prints
“?esk? republika”. The non-ASCII characters with diacritics are mangled.
The same problem exists in any locale: switch to an English locale in
this way, then search for “Abidjan”; it will be reported to be in “C?te
d’Ivoire”.

The issue appears to be that the LC_CTYPE category is not being changed.
According to locale(7):

> This category determines the interpretation of byte sequences as
> characters (e.g., single versus multibyte characters), character
> classifications (e.g., alphabetic or digit), and the behavior of
> character classes.

Rather than adding this category, just set all categories. It's true
that Initial Setup doesn't use, for instance, anything related to the
locale's measurement system (LC_MEASUREMENT, i.e. metric versus US
customary units) but it should be harmless to set them all.

The change in cc-common-language is probably unnecessary.
insert_language() is checking that the given locale actually exists on
the system before offering it to the user. Surely if
newlocale(LC_MESSAGES_MASK, x, 0) works then newlocale(LC_ALL_MASK, x,
0) will work? But there is no harm in using LC_ALL_MASK and the
consistency is clearer.

Fixes #177