keyboard: Correctly update labels for IBus engines
After the port from GTK 3 to 4, in general there's an additional level
of indirection: the children of priv->input_list are GtkListBoxRow
objects containing the widget whose user-data is the InputWidget. This
means we didn't find the InputWidget and therefore couldn't update its
display name.

Unfortunately, there is one exception to the rule that every child
of a GtkListBox is a GtkListBoxRow: the placeholder object, in our
case priv->no_results, is not wrapped in a GtkListBoxRow (see also
GNOME/gtk#4523). This means that walking the GtkWidget tree seems rather
fragile: it's difficult to tell whether each child of the GtkListBox
is a GtkListBoxRow, or the placeholder object priv->no_results, or
some third type of object that could be added by a future GTK version
(particularly since there is no particular type-safety here).

Instead of walking the widget tree, maintain our own parallel list of
known keyboard layouts and other input methods. For this list, we can
safely assert that every item is something that we put there, with the
invariant that it's a GtkBox with an InputWidget attached.

This means that we can reliably find the InputWidget, and update its
associated display name.

Fixes: ad500afc "keyboard: Port to GTK4"
Resolves: https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/issues/180
Signed-off-by: Simon McVittie <smcv@debian.org>