driver: Set a non-trivial window title
When running in "existing user" mode (which isn't done upstream, but
Debian and Ubuntu re-enable it because they don't have GNOME Tour),
the window title is visible in Alt+[key above Tab] and the Overview.
It defaults to "gnome-initial-setup", but looks better if we reuse the
translated name from the .desktop file.

Signed-off-by: Simon McVittie <smcv@debian.org>
(cherry picked from commit 2ed2453476bc62501896083109c542cf407c7ae1)