privacy: Move subtitles to separate labels
Prior to 71e3e2b4, the subtitles were contained into separate
GtkLabels connected to the the 'activate-link' signal. That commit
moved these labels to inside the corresponding AdwActionRows, but
did not account for the 'activate-link' signal.

To the best of my knowledgethere is currently no way to properly
support this with libadwaita rows. In addition to that, the latest
mockup [1] suggest placing these labels beneath the listbox anyway.

Move the subtitle labels to outside the action row, into a separate
labels beneath the listbox. Connect to the 'activate-link' signal
again.

Closes https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/issues/181

[1]
https://gitlab.gnome.org/Teams/Design/os-mockups/-/raw/master/initial-setup/initial-setup.png

(cherry picked from commit d230f18dc50bc40977b613845c8d50d1327cacf4)