privacy: Increase policy popover width
The [Mozilla Location Service privacy policy][mls] has a CSS breakpoint
at 760 wide, beyond which the logo and navigation are shown side-by-side
rather than stacked. Increasing the width of this popover slightly to be
wider than this breakpoint means that substantially more of the policy
can be read without scrolling, and subjectively looks better.

The current minimum size of 600×500 is already more than the
[HIG-specified][hig] minimum phone size of 360×294. 750×500 is within
the smallest recommended display size for GNOME, 1024×600.

[mls]: https://location.services.mozilla.com/privacy
[hig]: https://developer.gnome.org/hig/guidelines/adaptive.html