privacy: don't show ABRT switch if distro name is missing
GNOME does not run an ABRT server, so it does not make sense to use
"GNOME" as the fallback name if the distro name is missing from
/etc/os-release. In this case, let's just not show the ABRT option at
all, and keep it disabled.

This shouldn't affect much in practice, at least not today, because
currently only Fedora has an ABRT server, and if other distros were to
use ABRT in the future they would probably provide their name in
/etc/os-release. But might as well be careful in what we show to users
here.