assistant: Show Back button on summary page
commit f60b4622350468f7ef17f79d9bc6679bf8cce7b9 changed the
assistant to no longer show the back and forward buttons on the
last page.

Hiding the forward button makes sense: there's no more pages go
to.

Hiding the back button doesn't really make sense: there are a
bunch of pages the user could potentially want to revisit.

This commit shows the back button on the last page (either
the summary page or the install page).