goa: Suppress a possibly-uninitialised warning
The compiler cannot know that, in practice, one or other of the X11 or
Wayland code paths will initialise 'display'. Initialise it to NULL,
then assert after the two conditionally-compiled blocks that it is not
NULL. We can all enjoy hitting that assertion when a third display
server protocol is invented.