privacy: move web view functions
This just moves the existing functions higher in the file, so I can use
them lower in the file in a subsequent commit without adding forward
declarations. It's a separate commit because otherwise the diff of
actual changes would be hard to read.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/222>