Make WebKitGTK dependency optional
It's used to display the privacy policy for Mozilla Location Services.
That's nice to do, but not a strong enough reason for WebKitGTK to be a
hard requirement. Just remove the privacy policy link if it's not
present.

It's also used to display the distro privacy policy for ABRT, but Fedora
is the only distro that supports ABRT and Fedora no longer has a privacy
policy in /etc/os-release, so this code is unused in practice. Still,
it's low effort to keep the code working.

Fixes #203

Part-of: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/222>