ci: Switch back to Fedora
The switch to Arch was necessary when no other distro had webkit2gtk-5.0.

However, Arch is not workable as a base distro for stable branches of
Initial Setup because it is a rolling release, so the CI image becomes
un-buildable after some time. By contrast, so long as one sticks to a
particular Fedora release series, the CI image is likely to continue
building without changes well into the future.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/222>