Fix failure to log into user account
Currently gis_driver_get_user_permissions() is guaranteed to fail
because we call it after gis_driver_save_data(), which creates the user
account. We need to call it later at the right point.

Also, let's own the user_password that we store on the summary and
install page. Even if it's not expected to change, it doesn't seem very
safe to rely on the GisDriver not deleting it.