privacy: Always attach activate_link callback
Before bab50bd60e8103fcde4deb7f5f9809ea650a91f8 (“Make WebKitGTK
dependency optional”), a callback was always attached to
GtkLabel::activate-link, even if there were no links in the label.
That commit made the activate_link() callback be defined only if
WebKitGTK is enabled, and attached it to the label dynamically based on
the presence of both WebKitGTK support, and in the case of the ABRT
description whether a privacy policy URL is known.

As the previous commit shows, this is prone to programmer error.

Instead, always define the callback, with a no-op implementation which
always returns FALSE (“not handled”) if WebKitGTK is not available.
Attach the callback unconditionally when instantiating the .ui file.