privacy: Increase size of privacy policy window
Previously, the window was narrow enough to trigger a small-screen
breakpoint in the Mozilla Location Service privacy policy, which wraps
the navigation links at the top of the page to one per line, leaving
almost no space for the policy itself.

Increase the window size a little. Set the size with
GtkWindow:default-width and :default-height so that the window can
shrink below this size on extremely small screens.

(On small screens, 800×600 is the default size of the main window; on
large screens, 1024×768 is the fixed size.)