account: Use an existing string for error while caching
In 3bb784c6c93eea22b6c5679063fff64535445c14, cherry-picked from
1b0cb656c5025467dc6a6e86917be0b885751e78, a new translatable string is
introduced, breaking string freeze.

I think the bug fix is worth having. So, let's reuse the title for a
slightly different error, which is already marked for translation lower
in the file. This is still an improvement on the previous behaviour
where failing to cache the account would be silently ignored.

https://gitlab.gnome.org/Teams/Translation/Coordination/-/issues/100