common-languages: Avoid uninitialized variable usage
The function `gnome_parse_locale()` doesn't nullify the
variables passed to it if it can't parse correctly. This causes a
crash when passing some locales like "en".

Initialize the variables to NULL and early return with a false when
parsing fails to avoid this.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/merge_requests/240>