Fully suppress header onwards buttons if page has its own
This code was added in 69914dde but is not currently used. It is
intended to be used in future by pages which manage their own
'onwards' button. It seems incomplete - it only changes the
state of one of the three possible header buttons, and only its
visibility, not its sensitivity. It seems safer to ensure all
three possible header buttons are not visible and not sensitive.
We can conveniently do this by sending `set_navigation_button`
a widget which is none of the three buttons.

Signed-off-by: Adam Williamson <awilliam@redhat.com>