copy-worker: Rewrite to use plain glib
It's nice to copy the files with GFile since all the complexity is
hidden behind a single `g_file_copy` call. However, this brings in GVfs,
which makes it very difficult to start early in the session before D-Bus
is available.

Instead, use plan GLib/libc to copy the files. The semantics of what
metadata is copied should match what GFile was doing. This is hopefully
portable enough, but it does use a couple Unix specific APIs without the
full portability of GLib.

Fixes: #217