copy-worker: Always copy files
Rather than moving the files from the gnome-initial-setup homedir, copy
them. In practice the files will almost always be copied anyways since
the g-i-s user's homedir is on a tmpfs in `/run` and the target user's
homedir is most likely to be on real storage. Always making it a copy
means it will behave the same way in all scenarios including testing.
Moving the file also means the file is deleted from the g-i-s homedir,
which just causes a bunch of journal spam during initial setup.