Revert "password: Keep old password page UI if parental controls are disabled"
This reverts commit 7eae913a4076cbf6f3697e80158e3413ea71cab6:

> password: Keep old password page UI if parental controls are disabled
>
> While this means we are essentially maintaining three versions of the
> password page, it does mean that if parental controls are disabled,
> gnome-initial-setup doesn’t change its UI or translatable strings at
> all, i.e. doesn’t break UI freeze.
>
> This commit can be dropped in the next unstable development cycle.

This message was written in February 2020. The “next unstable
development cycle” it refers to was the cycle leading to GNOME 3.38.

The effect of this change is that, when setting up a system in the name
of Jane Bloggs, the password page for the user password (as opposed to
the parent password) will now always be titled:

> Set a Password for Jane Bloggs

regardless of whether Initial Setup was compiled with parental controls
support, and regardless of whether parental controls are being enabled.