password: Flag confirm entry if passwords don't match
Previously, the first password entry would have the red “error” state,
even though the message explaining the error – “The passwords do not
match” – is below the confirmation field.

Instead, mark the confirmation field as in error in that case. (Of
course we can never know which of the two fields is the one that has the
typo from the user's point of view.)