password: Don't mark empty password field as error
Previously, the primary password entry field would be considered to be
in error even if it is empty, and marked in red. This looks awkward
because the user has done nothing wrong.

The password strength meter, however, correctly treats the "no password"
case differently to the state where the password is weak.  Do the same
for the field itself. Eliminate the valid_password field of the struct,
which is never checked outside this function.

Fixes https://gitlab.gnome.org/GNOME/gnome-initial-setup/-/issues/162