password: Unify ::changed callbacks
Previously the password_changed() callback did not reset valid_confirm
before calling update_page_validation(), and update_page_validation()
checks valid_confirm and sets the page to complete or incomplete based
on its value.

There is still one residual difference between the handling of the
::changed signal between the two fields – typing in the primary password
field clears any error state in the confirmation field, but not
vice-versa – but I think this is clearer with an explicit condition in
a single callback, rather than two callbacks with one subtle difference.