Remove integration with keyring
We used to handle the keyring because gnome-initial-setup offered to set
up online accounts. This would store the login tokens in the keyring,
and then g-i-s would copy the keyring over to the real user session on
first login.

In GNOME 46 we removed the online accounts integration, so there's
nothing writing to the keyring anymore, and no more need to copy it.

Follow-up for 5ce2481938290e76a77f8ac8801cd23a58f634f6