Try run0 first if available instead of pkexec
pkexec is a SUID binary, which is discouraged these days as
they are a constant source of security issues since they elevate
privileges from an existing unprivileged context, under the control
of an unprivileged user.
run0 uses IPC to request the privilege elevation, and does not
require SUID programs, so try it first, and fallback to pkexec
if not found.